+++ torchkge/utils/datasets.py
@@ -18,7 +18,7 @@ from urllib.request import urlretrieve
 
 from torchkge.data_structures import KnowledgeGraph
 
-from torchkge.utils import get_data_home
+from torchkge.utils import get_data_home, safe_extract
 from torchkge.utils.operations import extend_dicts
 
 
@@ -307,28 +307,6 @@ def load_wikidatasets(which, limit_=0, data_home=None):
                     data_home + '/{}.tar.gz'.format(which))
 
         with tarfile.open(data_home + '/{}.tar.gz'.format(which), 'r') as tf:
-            
-            import os
-            
-            def is_within_directory(directory, target):
-                
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-            
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-                
-                return prefix == abs_directory
-            
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-            
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-            
-                tar.extractall(path, members, numeric_owner=numeric_owner) 
-                
-            
             safe_extract(tf, data_home)
         remove(data_home + '/{}.tar.gz'.format(which))
