+++ torchkge/utils/modeling.py
@@ -11,7 +11,7 @@ from torch.nn.init import xavier_uniform_
 import pickle
 import tarfile
 
-from .data import get_data_home
+from torchkge.utils import get_data_home, safe_extract
 
 from os import makedirs, remove
 from os.path import exists
@@ -41,28 +41,6 @@ def load_embeddings(model, dim, dataset, data_home=None):
         urlretrieve("https://graphs.telecom-paris.fr/data/torchkge/models/{}_{}_{}.tar.gz".format(model, dataset, dim),
                     targz_file)
         with tarfile.open(targz_file, 'r') as tf:
-            
-            import os
-            
-            def is_within_directory(directory, target):
-                
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-            
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-                
-                return prefix == abs_directory
-            
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-            
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-            
-                tar.extractall(path, members, numeric_owner=numeric_owner) 
-                
-            
             safe_extract(tf, data_path)
         remove(targz_file)
